+++ core/CHANGELOG.md
@@ -8,6 +8,7 @@ _YYYY-MM-DD_
 
 * Dependency loop error now lists all test cases that formed a cycle
 * `foldGroup` now takes `[b]` instead of `b` as its last argument to allow for custom fold strategies. This is a backwards incompatible change, but you can get the old behavior by applying `mconcat`.
+* Dependencies can now be defined pattern-free with `sequentialTestGroup`. As a consequence, `TestGroup` now has an additional field `ExecutionMode`, specifying whether a test group should run in parallel or sequentially. (#343)
 
 Version 1.4.3
 ---------------